调整字段名,数据类型

+++ src/main/java/com/snow/storeapi/entity/VipInfo.java
@@ -45,9 +45,9 @@ public class VipInfo implements Serializable {
 
     private String certNo;
 
-    private Integer birthDiscount;
+    private BigDecimal birthDiscount;
 
-    private Integer vipDiscount;
+    private BigDecimal vipDiscount;
 
     private BigDecimal balance;
